requirements.txt
dplib-py==1.0.0
requests==2.32.3